+++ lib/assets/core/javascripts/cartodb3/editor/layers/layer-content-views/analyses/analysis-buttons-view.js
