+++ lib/assets/core/javascripts/cartodb3/editor/layers/layer-content-views/analyses/analysis-controls-view.js
@@ -12,7 +12,7 @@ var AnalysesQuotaProvider = require('./analyses-quota/analyses-quota-provider');
 var AnalysesQuotaEstimation = require('./analyses-quota/analyses-quota-estimation-input');
 var AnalysesQuotaEnough = require('./analyses-quota/analyses-quota-enough');
 var checkAndBuildOpts = require('../../../../helpers/required-opts');
-var AnalysisButtonView = require('./analysis-button-view');
+var AnalysisButtonsView = require('./analysis-buttons-view');
 var AnalysisQuotaView = require('./analysis-quota-view');
 
 var REQUIRED_OPTS = [
@@ -27,7 +27,7 @@ var REQUIRED_OPTS = [
 ];
 
 /**
- * View representing the apply button for a form
+ * View representing the apply and cancel/delete buttons for a form
  */
 module.exports = CoreView.extend({
   className: 'Options-bar u-flex u-justifySpace',
@@ -37,7 +37,7 @@ module.exports = CoreView.extend({
     'click .js-delete': '_onDeleteClicked'
   },
 
-  _buttonView: null,
+  _buttonsView: null,
 
   initialize: function (opts) {
     checkAndBuildOpts(opts, REQUIRED_OPTS, this);
@@ -81,11 +81,11 @@ module.exports = CoreView.extend({
     var view;
 
     if (this._isAnalysisDone() && !this._hasChanges()) {
-      view = this._createButtonView();
+      view = this._createButtonsView();
     } else {
       // If the analysis doesn't require quota, let's not wait or check the fetching state
       if (!requiresQuota || !isValid) {
-        view = this._createButtonView();
+        view = this._createButtonsView();
       } else {
         view = this._createQuotaView();
       }
@@ -159,7 +159,7 @@ module.exports = CoreView.extend({
     });
   },
 
-  _createButtonView: function () {
+  _createButtonsView: function () {
     this._viewModel.set({
       labelSave: _t('editor.layers.analysis-form.apply-btn'),
       labelDelete: !this._analysisNode
@@ -168,14 +168,14 @@ module.exports = CoreView.extend({
       isDisabled: !this._canSave()
     });
 
-    if (this._buttonView === null) {
-      this._buttonView = new AnalysisButtonView({
+    if (this._buttonsView === null) {
+      this._buttonsView = new AnalysisButtonsView({
         template: template,
         model: this._viewModel
       });
     }
 
-    return this._buttonView;
+    return this._buttonsView;
   },
 
   _buildQuotaViewInfo: function () {
@@ -237,7 +237,7 @@ module.exports = CoreView.extend({
           .then(checkEnoughQuota, errorCallback)
           .then(buildPayload, errorCallback);
       } else {
-        // to render the simple button view.
+        // to render the simple buttons view.
         this.render();
       }
     } else if (!dsReady) {
